Module obddimal::core

Modules

apply

The Apply operator

bdd_manager

All BDD building and manipulation functionality

bdd_node

Module containing type definitions for the elements of the BDD: Nodes (DDNode, NodeID) and Variables (VarID)

dvo

Implementation of dynamic variable ordering techniques

ite

The ITE operator

modify_var_nodes

An operation allowing to modify all nodes belonging to specific variables/layers.

options

Options for BDD building

order 🔒

Utility functions related to variable order

reduce 🔒

BDD reduction

slicing

BDD slicing

swap 🔒

Implementation of BDD layer swap

test 🔒