Module obddimal::core::bdd_node

Module containing type definitions for the elements of the BDD: Nodes (DDNode, NodeID) and Variables (VarID)

Structs

DDNode

Element of a BDD. Note that the Node contains its own ID. This may be set to zero until it has been assigned, and most importantly is not considered in hashing and equality testing.

NodeID

VarID

Constants

ONE

Terminal node “one”

ZERO

Terminal node “zero”