Module obddimal::core::dvo::dvo_schedules

Strategies for when and how to run DVO during BDD generation

Structs

AlwaysOnce

Run one iteration of sifting for all variables, every time it’s called. See DDManager::sift_single_var() for max_increase parameter.

AlwaysUntilConvergence

Always perform sifting of all variables until the number of nodes does not change anymore.

AtThreshold

Call the underlying strategy if the node count exceeds the specified limit

NoDVOSchedule

Dummy DVO implementation that does nothing

SiftingAtThreshold

Performs sifting until the number of nodes does not change anymore, but only if the initial number of nodes exceeds a configurable threshold.

TimeSizeLimit

Calls the underlying DVO mode if the specified duration has passed since the last invocation, or the nodes table exceeds the size specified in limit.

Enums

DVOScheduleEnum

This contains all available DVO implementations

Traits

DVOSchedule 🔒

Implements run_dvo()