Struct obddimal::core::dvo::dvo_schedules::SiftingAtThreshold

pub struct SiftingAtThreshold {
    underlying: AtThreshold,
}

Performs sifting until the number of nodes does not change anymore, but only if the initial number of nodes exceeds a configurable threshold.

Fields
underlying: AtThreshold
Implementations

impl SiftingAtThreshold

pub fn new(active_nodes_threshold: usize) -> SiftingAtThreshold