obddimal::core::dvo::dvo_schedules
Trait DVOSchedule

pub(crate) trait DVOSchedule {
    // Required method
    fn run_dvo(
        &mut self,
        num_clause: usize,
        man: &mut DDManager,
        f: NodeID,
        bar: &Option<ProgressBar>,
    ) -> NodeID;
}

Implements run_dvo()

Required Methods

fn run_dvo( &mut self, num_clause: usize, man: &mut DDManager, f: NodeID, bar: &Option<ProgressBar>, ) -> NodeID

This gets called after a CNF clause has been integrated. The current root node is f, the implementation must return the new root node ID, even if it does not change.

num_clause: The index of the current clause, in integration order (which may differ from the order defined in the input CNF).