Struct obddimal::views::bdd_view::BddView

pub struct BddView {
    man: Arc<RwLock<DDManager>>,
    man_id: usize,
    root: NodeID,
    sliced_vars: HashSet<VarID>,
}

Fields
man: Arc<RwLock<DDManager>>man_id: usizeroot: NodeIDsliced_vars: HashSet<VarID>
Implementations

impl BddView

pub(crate) fn new(root: NodeID, manager: Arc<RwLock<DDManager>>) -> Arc<Self>

pub(crate) fn new_with_sliced( root: NodeID, manager: Arc<RwLock<DDManager>>, sliced_vars: HashSet<VarID> ) -> Arc<Self>

pub fn get_manager(&self) -> Arc<RwLock<DDManager>>

Gives access to the manager which stores the BDD.

pub fn get_root(&self) -> NodeID

Returns the NodeID of the root node of this BDD.

pub fn get_sliced_variables(&self) -> HashSet<VarID>

Returns the VarIDs of the sliced variables of this BDD.

pub fn zero(manager: Arc<RwLock<DDManager>>) -> Arc<BddView>

Returns a BddView for a BDD which represents the function which is constant 0.

pub fn one(manager: Arc<RwLock<DDManager>>) -> Arc<BddView>

Returns a BddView for a BDD which represents the function which is constant 1.

pub fn from_dimacs( dimacs: String, order: Option<Vec<usize>>, options: Options ) -> Result<Arc<BddView>, String>

Build a bdd from dimacs. The BDD is stored in a new DDManager created by this function.

pub fn xor_prim( manager: Arc<RwLock<DDManager>>, without_vars: &HashSet<VarID> ) -> Arc<BddView>

Builds a xor ladder with all variables in the DDManager except the given ones and returns a BddView for the resulting BDD.

pub fn not(&self) -> Arc<Self>

Returns a (new) view on a BDD which represents the inverse of the function of the BDD.

pub fn and(&self, other: &Self) -> Arc<Self>

Returns a (new) view on the BDD resulting from connecting this views’ BDD and another one given via the parameter with an and.

pub fn or(&self, other: &Self) -> Arc<Self>

Returns a (new) view on the BDD resulting from connecting this views’ BDD and another one given via the parameter with an or.

pub fn xor(&self, other: &Self) -> Arc<Self>

Returns a (new) view on the BDD resulting from connecting this views’ BDD and another one given via the parameter with an xor.

pub fn create_slice(&self, keep: &HashSet<VarID>) -> Arc<Self>

Creates a slice of the BDD containing only the given variables.

keep - The variables to keep

pub fn create_slice_without_vars(&self, remove: &HashSet<VarID>) -> Arc<Self>

Creates a slice of the BDD containing all except the given variables.

remove - The variables to remove

fn nodelist_to_viewlist(nodes: Vec<NodeID>, man: DDManager) -> Vec<Arc<BddView>>

pub fn load_from_dddmp_file( filename: String ) -> Result<Vec<Arc<BddView>>, String>

Loads the BDDs from a .dddmp file.

filename - Name of the .dddmp file.

pub fn is_sat(&self) -> bool

Returns, whether the function represented by this BDD is satisfyable.

pub fn sat_count(&self) -> BigUint

Returns the #SAT result for the function represented by this BDD.

pub fn graphviz(&self) -> String

Generate graphviz for the BDD.