Struct obddimal::core::bdd_manager::DDManager

pub struct DDManager {
    id: usize,
    pub nodes: HashMap<NodeID, DDNode>,
    pub(crate) var2level: Vec<usize>,
    pub(crate) level2nodes: Vec<HashSet<DDNode>>,
    pub(super) ite_c_table: HashMap<(NodeID, NodeID, NodeID), NodeID>,
    pub(super) apply_c_table: HashMap<(ApplyOperation, NodeID, NodeID), NodeID>,
    views: WeakHashSet<Weak<BddView>>,
}

Container combining the nodes list, unique tables, information about current variable ordering and computed table.

Fields
id: usize

ID to identify the DDManager

nodes: HashMap<NodeID, DDNode>

Node List

var2level: Vec<usize>

Variable ordering: var2level[v.0] is the depth of variable v in the tree

level2nodes: Vec<HashSet<DDNode>>

Unique Table for each variable. Note that the indices into this table are the depths of the variables, not their IDs. The depth of a variable can be determined through var2level!

ite_c_table: HashMap<(NodeID, NodeID, NodeID), NodeID>

Computed Table for ITE: maps (f,g,h) to ite(f,g,h)

apply_c_table: HashMap<(ApplyOperation, NodeID, NodeID), NodeID>

Computed Table for Apply: maps (op,u,v) to apply(op,u,v)

views: WeakHashSet<Weak<BddView>>

Set of Views for BDDs in this manager

Implementations

impl DDManager

pub fn from_instance( instance: &mut Instance, order: Option<Vec<usize>>, options: Options ) -> Result<(DDManager, NodeID), String>

Builds a BDD from a CNF read from DIMACS.

instance - Input CNF

order - Optional initial variable ordering, see crate::misc::static_ordering for implementations

options - DVO and progress bar settings

TODO Variant that works on already existing DDManager…

let mut instance = dimacs::parse_dimacs(
    &fs::read_to_string("examples/trivial.dimacs")
        .expect("Failed to read dimacs file."),
)
.expect("Failed to parse dimacs file.");
let order = Some(static_ordering::force(&instance));
let dvo = dvo_schedules::SiftingAtThreshold::new(5);
let (man, bdd) = DDManager::from_instance(
    &mut instance,
    order,
    Options::default().with_progressbars().with_dvo(dvo.into()),
).unwrap();

impl DDManager

pub fn apply(&mut self, op: ApplyOperation, u: NodeID, v: NodeID) -> NodeID

impl DDManager

fn bootstrap(&mut self)

Initialize the BDD with zero and one constant nodes

pub(crate) fn get_or_add_view(&mut self, view: BddView) -> Arc<BddView>

Add a view to the BDD Manager

pub fn get_views(&self) -> Vec<Arc<BddView>>

Get all views for this BDD Manager

pub fn get_roots(&self) -> Vec<NodeID>

Get all root nodes for this BDD Manager

pub(crate) fn get_id(&self) -> usize

Get the id of the BDD Manager (mostly relevant for checking for equality of two BDD Managers).

pub(crate) fn prepare_varorder(&mut self, varorder: Vec<usize>)

Initializes the BDD for a specific variable ordering.

Panics

Only allowed on empty DDManagers. If called on a non-empty DDManager, this function will panic!

fn ensure_order(&mut self, target: VarID)

Ensure var2level vec is valid up to specified variable

fn add_node(&mut self, node: DDNode) -> NodeID

Insert Node. ID is assigned for nonterminal nodes (var != 0). This does not check the unique table, you should do so before using!

pub(crate) fn node_get_or_create(&mut self, node: &DDNode) -> NodeID

Search for Node, create if it doesnt exist

pub(crate) fn zero(&self) -> NodeID

pub(crate) fn one(&self) -> NodeID

pub fn ith_var(&mut self, var: VarID) -> NodeID

pub fn nith_var(&mut self, var: VarID) -> NodeID

pub fn not(&mut self, f: NodeID) -> NodeID

pub fn exists(&mut self, f: NodeID, vars: &HashSet<VarID>) -> NodeID

pub fn forall(&mut self, f: NodeID, vars: &HashSet<VarID>) -> NodeID

pub fn exists_multiple( &mut self, fs: &[NodeID], vars: &HashSet<VarID> ) -> HashMap<NodeID, NodeID>

Existential quantification, but on multiple BDDs at the same time. Returns a HashMap, which can be used to translate the root nodes of the old BDDs to the root nodes of the resulting BDDs.

pub fn forall_multiple( &mut self, fs: &[NodeID], vars: &HashSet<VarID> ) -> HashMap<NodeID, NodeID>

Universal quantification, but on multiple BDDs at the same time. Returns a HashMap, which can be used to translate the root nodes of the old BDDs to the root nodes of the resulting BDDs.

pub fn and(&mut self, f: NodeID, g: NodeID) -> NodeID

pub fn or(&mut self, f: NodeID, g: NodeID) -> NodeID

pub fn xor(&mut self, f: NodeID, g: NodeID) -> NodeID

pub(super) fn min_by_order( &self, fvar: VarID, gvar: VarID, hvar: VarID ) -> VarID

Find top variable: Highest in tree according to order

pub fn xor_prim(&mut self, without_vars: &HashSet<VarID>) -> NodeID

Creates an XOR “ladder”

pub fn verify(&self, f: NodeID, trues: &[usize]) -> bool

pub fn get_reachable(&self, roots: &[NodeID]) -> HashSet<NodeID>

Returns a set containing all nodes reachable from the given root nodes.

pub fn purge_retain(&mut self, root: NodeID)

Removes nodes which do not belong to the BDD with the specified root node from the DDManager.

root - The root node of the BDD which should remain in the DDManager.

pub fn purge_retain_multi(&mut self, roots: &[NodeID])

Removes nodes which do not belong to any of the BDDs with the specified root nodes from the DDManager.

roots - The root nodes of the BDDs which should remain in the DDManager.

pub fn clean(&mut self)

Removes nodes which do not belong to any of the BDDs for which views exist from the DDManager.

pub fn clear_c_table(&mut self)

impl DDManager

fn var_at_level(&self, level: usize) -> Option<VarID>

Find the variable at specified level

fn sift_single_var( &mut self, var: VarID, max_increase: Option<usize>, f: NodeID ) -> NodeID

Swap layer containing specified variable first to the bottom of the BDD, then to the top, and then to the position which resulted in smallest BDD size. Optional parameter max_increase stops swapping in either direction once the difference between BDD size and current optimum exceeds threshold.

pub fn sift_all_vars( &mut self, f: NodeID, progressbar: bool, max_increase: Option<usize> ) -> NodeID

Perform sifting for every layer containing at least one variable. If progressbar is true, display a progress bar in the terminal which shows the number of layers already processed. See Self::sift_single_var() for max_increase parameter.

impl DDManager

pub fn ite(&mut self, f: NodeID, g: NodeID, h: NodeID) -> NodeID

impl DDManager

pub fn modify_var_nodes( &mut self, relevant_nodes: &HashSet<NodeID>, vars: &HashSet<VarID>, func: &mut NodeModFunction ) -> HashMap<NodeID, NodeID>

Modifies all nodes in the given BDD(s) corresponding to the given variables using the given function. Also corrects the references in nodes above the modified nodes so the resulting BDD(s) are still valid. The “modification” thereby creates new nodes, so the original BDDs can still be used later.

relevant_nodes - A set containing all nodes of the BDDs for which this operation is supposed to be run.

vars - The variables, for which the nodes should be “modified”.

func - The function which should be applied to the nodes corresponding to the given variables.

Returns

A HashMap which can be used to translate the root nodes of the original BDDs to the root nodes of the new (“modified”) BDDs.

Notes

May create temporary nodes, consider using purge_retain_multi or purge_retain afterwards.

impl DDManager

pub(crate) fn reduce(&mut self, v: NodeID) -> NodeID

Reduces the BDD. This changes Node IDs, the new Node ID of the function passed to the function is returned. All other known Node IDs are invalid after reduce. Note: Reduction does not imply removing nodes not part of the specified function. Other functions may be present in the tree (multiple root nodes, or root above function).

See also “Graph-Based Algorithms for Boolean Function Manipulation” by Bryant (10.1109/TC.1986.1676819)

impl DDManager

pub fn create_slice(&mut self, root: NodeID, keep: &HashSet<VarID>) -> NodeID

Creates a slice of a BDD containing only the given variables.

keep - The variables to keep

root - The root node of the BDD that is supposed to be sliced

Returns

The id of the root node of the slice.

Notes

May create temporary nodes, consider using purge_retain_multi or purge_retain afterwards.

pub fn create_slice_without_vars( &mut self, root: NodeID, remove: &HashSet<VarID> ) -> NodeID

Creates a slice of a BDD containing all except the given variables.

remove - The variables to remove

root - The root node of the BDD that is supposed to be sliced

Returns

The id of the root node of the slice.

Notes

May create temporary nodes, consider using purge_retain_multi or purge_retain afterwards.

pub fn create_slices( &mut self, roots: &[NodeID], keep: &HashSet<VarID> ) -> HashMap<NodeID, NodeID>

Creates slices of BDDs containing only the given variables.

keep - The variables to keep

roots - The root nodes of the BDDs that are supposed to be sliced

Returns

A HashMap giving the id of the root node of the created slice for each BDD (the keys to the HashMap are the ids of the root nodes of the original BDDs, as given in the node_ids parameter).

Notes

May create temporary nodes, consider using purge_retain_multi or purge_retain afterwards.

pub fn create_slices_without_vars( &mut self, roots: &[NodeID], remove: &HashSet<VarID> ) -> HashMap<NodeID, NodeID>

Creates slices of BDDs containing all except the given variables.

remove - The variables to remove

roots - The root nodes of the BDDs that are supposed to be sliced

Returns

A HashMap giving the id of the root node of the created slice for each BDD (the keys to the HashMap are the ids of the root nodes of the original BDDs, as given in the node_ids parameter).

Notes

May create temporary nodes, consider using purge_retain_multi or purge_retain afterwards.

impl DDManager

pub fn swap(&mut self, a: VarID, b: VarID, f: NodeID) -> NodeID

Swaps graph layers of variables a and b. Requires a to be directly above b or vice versa. Performs reduction which may change NodeIDs. Returns new NodeID of f.

impl DDManager

pub fn load_from_bdd_file_with_statistics( filename: String ) -> LoadResultWithStatistics

Reads a (multi-rooted) BDD and corresponding statistics from a .bdd file.

filename - Name of the .bdd file.

let (man, bdds, views, statistics) = DDManager::load_from_bdd_file_with_statistics("examples/simple.bdd".to_string()).unwrap();

pub fn load_from_bdd_file(filename: String) -> LoadResult

Reads a (multi-rooted) BDD from a .bdd file.

filename - Name of the .bdd file.

let (man, bdds, views) = DDManager::load_from_bdd_file("examples/simple.bdd".to_string()).unwrap();

pub fn write_all_views_to_bdd_file( &self, filename: String, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<(), String>

Writes a (multi-rooted) BDD and all its views to a .bdd file.

self - The DDManager containing the BDD.

filename - Name of the .bdd file.

statistics - Optional HashMap containing statistics about individual nodes, see NodeStatistics.

pub fn write_to_bdd_file( &self, filename: String, roots: Vec<NodeID>, views: Option<Vec<Arc<BddView>>>, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<(), String>

Writes a (multi-rooted) BDD to a .bdd file.

self - The DDManager containing the BDD.

filename - Name of the .bdd file.

roots - The roots of the BDD.

views - Optionally the views to store for the BDDs.

statistics - Optional HashMap containing statistics about individual nodes, see [NodeStatistics].

fn generate_bdd_file_string( &self, roots: Vec<NodeID>, views: Option<Vec<Arc<BddView>>>, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<String, String>

impl DDManager

pub(super) fn load_from_bdd_file_object_with_statistics( bdd_file: BddFile ) -> LoadResultWithStatistics

Reads a (multi-rooted) BDD and corresponding statistics from a bdd file object.

bdd_file - The BddFile object.

pub(super) fn generate_bdd_file_object( &self, roots: Vec<NodeID>, views: Option<Vec<Arc<BddView>>>, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> BddFile

pub(super) fn load_bdd_from_nodelist( self, nodes: HashMap<NodeID, (VarID, NodeID, NodeID)>, varorder: Vec<usize>, roots: Vec<NodeID>, terminals: (NodeID, NodeID) ) -> (DDManager, Vec<NodeID>)

Loads a BDD from a Nodelist (containing all nodes from a BDD) into the DDManager.

Panics

Only allowed on empty DDManagers. If called on a non-empty DDManager, this function will panic!

pub(super) fn load_bdd_from_nodelist_with_translation( self, nodes: HashMap<NodeID, (VarID, NodeID, NodeID)>, varorder: Vec<usize>, roots: Vec<NodeID>, terminals: (NodeID, NodeID) ) -> (DDManager, Vec<NodeID>, HashMap<NodeID, NodeID>)

Loads a BDD from a Nodelist (containing all nodes from a BDD) into the DDManager.

Panics

Only allowed on empty DDManagers. If called on a non-empty DDManager, this function will panic!

impl DDManager

pub fn load_from_dddmp_file( filename: String ) -> Result<(DDManager, Vec<NodeID>), String>

Reads a (multi-rooted) BDD from a .dddmp file.

filename - Name of the .dddmp file.

let (man, bdds) =
DDManager::load_from_dddmp_file("examples/sandwich.dimacs.dddmp".to_string()).unwrap();

fn parse_bcdd_from_dddmp_file(file: File) -> Result<Bcdd, String>

Parses a BCDD from a .dddmp file.

file - The file to be parsed

fn parse_bcdd_nodelist<I>( lines: &mut Peekable<I>, expected_nodecount: usize ) -> Result<(HashMap<isize, (VarID, isize, isize)>, isize), String>
where I: Iterator<Item = String>,

Parses the nodelist of a BCDD from a .dddmp file.

lines - An iterator over the lines of the file, the header including the .nodes mark should already be consumed

expected_nodecount - The number of nodes that is expected to be parsed (used for sanity checks)

fn convert_bcdd_to_bdd(bcdd: &Bcdd) -> (DDManager, Vec<NodeID>)

Converts a BCDD to a normal BDD

bcdd - The BCDD to be converted

fn create_bcdd_node_parent_information( bcdd: &Bcdd ) -> HashMap<isize, (HashSet<ParentNode>, HashSet<ParentNode>)>

Creates a HashMap containing information about the parents of each node and the edges connecting them.

bcdd - The BCDD containing the nodes

fn create_bcdd_layer_to_nodes(bcdd: &Bcdd) -> HashMap<usize, HashSet<isize>>

Creates a HashMap containing the node IDs for each layer.

bcdd - The BCDD containing the nodes

fn convert_bcdd_to_bdd_nodes( node_parent_information: HashMap<isize, (HashSet<ParentNode>, HashSet<ParentNode>)>, layer_to_nodes: HashMap<usize, HashSet<isize>>, bcdd: &Bcdd ) -> HashMap<isize, (VarID, isize, isize)>

Creates a HashTable containing the nodes of a BDD representing the same function as the given BCDD.

node_parent_information - HashMap storing the parents of a node.

layer_to_nodes - HashMap mapping the layers of the bcdd to its nodes

bcdd - The BCDD which is going to be converted to a BDD.

impl DDManager

pub fn load_from_bdd_file_json_with_statistics( filename: String ) -> LoadResultWithStatistics

Reads a (multi-rooted) BDD and corresponding statistics from a .bdd.json file.

filename - Name of the .bdd.json file.

pub fn load_from_bdd_file_json(filename: String) -> LoadResult

Reads a (multi-rooted) BDD from a .bdd.json file.

filename - Name of the .bdd.json file.

pub fn write_all_views_to_bdd_file_json( &self, filename: String, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<(), String>

Writes a (multi-rooted) BDD and all its views to a .bdd.json file.

self - The DDManager containing the BDD.

filename - Name of the .bdd.json file.

statistics - Optional HashMap containing statistics about individual nodes, see NodeStatistics.

pub fn write_to_bdd_file_json( &self, filename: String, roots: Vec<NodeID>, views: Option<Vec<Arc<BddView>>>, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<(), String>

Writes a (multi-rooted) BDD to a .bdd.json file.

self - The DDManager containing the BDD.

filename - Name of the .bdd.json file.

roots - The roots of the BDD.

views - Optionally the views to store for the BDDs.

statistics - Optional HashMap containing statistics about individual nodes, see NodeStatistics.

fn generate_bdd_file_json_string( &self, roots: Vec<NodeID>, views: Option<Vec<Arc<BddView>>>, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<String, String>

impl DDManager

pub fn load_from_bdd_file_xml_with_statistics( filename: String ) -> LoadResultWithStatistics

Reads a (multi-rooted) BDD and corresponding statistics from a .bdd.xml file.

filename - Name of the .bdd.xml file.

pub fn load_from_bdd_file_xml(filename: String) -> LoadResult

Reads a (multi-rooted) BDD from a .bdd.xml file.

filename - Name of the .bdd.xml file.

pub fn write_all_views_to_bdd_file_xml( &self, filename: String, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<(), String>

Writes a (multi-rooted) BDD and all its views to a .bdd.xml file.

self - The DDManager containing the BDD.

filename - Name of the .bdd.xml file.

statistics - Optional HashMap containing statistics about individual nodes, see NodeStatistics.

pub fn write_to_bdd_file_xml( &self, filename: String, roots: Vec<NodeID>, views: Option<Vec<Arc<BddView>>>, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<(), String>

Writes a (multi-rooted) BDD to a .bdd.xml file.

self - The DDManager containing the BDD.

filename - Name of the .bdd.xml file.

roots - The roots of the BDD.

views - Optionally the views to store for the BDDs.

statistics - Optional HashMap containing statistics about individual nodes, see NodeStatistics.

fn generate_bdd_file_xml_string( &self, roots: Vec<NodeID>, views: Option<Vec<Arc<BddView>>>, statistics: Option<HashMap<NodeID, NodeStatistics>> ) -> Result<String, String>

impl DDManager

pub fn graphviz(&self, f: NodeID) -> String

Generate graphviz for the provided function, not including any graph nodes not part of the function. TODO: Graphviz of all functions in DDManager

impl DDManager

pub fn is_sat(&self, node: NodeID) -> bool

pub fn sat_count(&self, f: NodeID) -> BigUint

fn sat_count_rec( &self, f: NodeID, cache: &mut HashMap<NodeID, BigUint> ) -> BigUint

pub fn count_active(&self, f: NodeID) -> usize