obddimal::views::bdd_view
Struct BddView

pub struct BddView {
    man: Arc<RwLock<DDManager>>,
    man_id: usize,
    root: NodeID,
    removed_vars: HashSet<VarID>,
    atomic_sets: Option<HashMap<VarID, HashSet<VarID>>>,
}

Fields
man: Arc<RwLock<DDManager>>man_id: usizeroot: NodeIDremoved_vars: HashSet<VarID>atomic_sets: Option<HashMap<VarID, HashSet<VarID>>>
Implementations

impl BddView

pub(crate) fn new(root: NodeID, manager: Arc<RwLock<DDManager>>) -> Arc<Self>

pub(crate) fn new_with_removed_vars( root: NodeID, manager: Arc<RwLock<DDManager>>, removed_vars: HashSet<VarID>, ) -> Arc<Self>

fn new_with_atomic_sets( root: NodeID, manager: Arc<RwLock<DDManager>>, removed_vars: HashSet<VarID>, atomic_sets: Option<HashMap<VarID, HashSet<VarID>>>, ) -> Arc<Self>

pub fn get_manager(&self) -> Arc<RwLock<DDManager>>

Gives access to the manager which stores the BDD.

pub fn get_root(&self) -> NodeID

Returns the NodeID of the root node of this BDD.

pub fn get_removed_variables(&self) -> HashSet<VarID>

Returns the VarIDs of the removed variables of this BDD.

pub fn get_optimizations(&self) -> Option<HashMap<VarID, HashSet<VarID>>>

Returns a HashMap containing for each atomic set that was used for optimization the variable that represents the atomic set (as the key) and the set of variables that got removed during the optimization (as the Values) or None if no optimizations were applied.

pub fn zero(manager: Arc<RwLock<DDManager>>) -> Arc<BddView>

Returns a BddView for a BDD which represents the function which is constant 0.

pub fn one(manager: Arc<RwLock<DDManager>>) -> Arc<BddView>

Returns a BddView for a BDD which represents the function which is constant 1.

pub fn from_dimacs( dimacs: String, order: Option<Vec<usize>>, options: Options, ) -> Result<Arc<BddView>, String>

Build a bdd from dimacs. The BDD is stored in a new DDManager created by this function.

pub fn xor_prim( manager: Arc<RwLock<DDManager>>, without_vars: &HashSet<VarID>, ) -> Arc<BddView>

Builds a xor ladder with all variables in the DDManager except the given ones and returns a BddView for the resulting BDD.

pub fn not(&self) -> Arc<Self>

Returns a (new) view on a BDD which represents the inverse of the function of the BDD.

pub fn exists(&self, vars: &HashSet<VarID>) -> Arc<Self>

Returns a (new) view on the BDD resulting from applying exists with the given variables to the BDD.

pub fn forall(&self, vars: &HashSet<VarID>) -> Arc<Self>

Returns a (new) view on the BDD resulting from applying forall with the given variables to the BDD.

pub fn and(&self, other: &Self) -> Arc<Self>

Returns a (new) view on the BDD resulting from connecting this views’ BDD and another one given via the parameter with an and.

pub fn or(&self, other: &Self) -> Arc<Self>

Returns a (new) view on the BDD resulting from connecting this views’ BDD and another one given via the parameter with an or.

pub fn xor(&self, other: &Self) -> Arc<Self>

Returns a (new) view on the BDD resulting from connecting this views’ BDD and another one given via the parameter with an xor.

pub fn create_slice(&self, keep: &HashSet<VarID>) -> Arc<Self>

Creates a slice of the BDD containing only the given variables.

keep - The variables to keep

pub fn create_slice_without_vars(&self, remove: &HashSet<VarID>) -> Arc<Self>

Creates a slice of the BDD containing all except the given variables.

remove - The variables to remove

fn nodelist_to_viewlist(nodes: Vec<NodeID>, man: DDManager) -> Vec<Arc<BddView>>

pub fn load_from_dddmp_file( filename: String, ) -> Result<Vec<Arc<BddView>>, String>

Loads the BDDs from a .dddmp file.

filename - Name of the .dddmp file.

pub fn is_sat(&self) -> bool

Returns, whether the function represented by this BDD is satisfyable.

pub fn sat_count(&self) -> Natural

Returns the #SAT result for the function represented by this BDD.

fn count_models_by_variable(&self) -> (Natural, HashMap<VarID, Natural>)

pub fn identify_atomic_sets(&self) -> Vec<HashSet<VarID>>

pub fn optimize_through_atomic_sets(&self) -> Option<Arc<Self>>

Optimizes the BDD by making use of Atomic Sets. Returns None if this view already uses an optimized BDD or otherwise a (new) View for the optimized BDD.

pub fn graphviz(&self) -> String

Generate graphviz for the BDD.

pub fn evaluate(&self, trues: &[VarID]) -> bool

Evaluates the BDD, setting all given vars to 1 and all remaining vars to 0.

trues - The variables that should be set to true

pub fn count_nodes(&self) -> usize

Counts how many nodes the BDD consists of.